Fix event DTO type and structure mismatch between warehouse and inventory services

+++ inventory-service/src/main/java/com/core/inventoryservice/dto/ConfirmedOrderDTO.java
@@ -1,6 +1,5 @@
 package com.core.inventoryservice.dto;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 import com.core.inventoryservice.domain.OrderStatus;
@@ -15,6 +14,6 @@ public record ConfirmedOrderDTO(
 		String customerPhone,
 		String customerAddress,
 		OrderStatus orderCurrentStatus,
-		BigDecimal totalAmount,
+		Double totalAmount,
 		List<@Valid ProductDTO> products
 ) {}
