+++ inventory-service/src/main/java/com/core/inventoryservice/dto/OrderDTO.java
@@ -1,6 +1,5 @@
 package com.core.inventoryservice.dto;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 import com.core.inventoryservice.domain.OrderStatus;
@@ -13,7 +12,7 @@ public record OrderDTO(
 		String customerPhone,
 		String customerAddress,
 		OrderStatus orderCurrentStatus,
-		BigDecimal totalAmount,
+		Double totalAmount,
 		List<@Valid ItemDTO> items,
 		List<OrderStatusStateDTO> statusHistory
 ) {}
